requirements.txt
matplotlib
pandas
asyncio
bleak
pyserial
simplekml
sqlalchemy
psycopg2